backend/pyproject.toml
[tool.poetry]
name = "notepad-backend"
version = "0.1.0"
description = "Backend service for the notepad application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 